Generalize meme generation interface
We can now support multiple boxes, and arbitrary positions via drag
and drop.

Still missing: the ability to save and load a layout for future use.

The javascript code most likely shows I had little idea of what I was
doing, but represents my understanding of the problem area at the
moment, which is extremely limited.